third_party/blink/web_tests/fast/dom/Window/atob-btoa-expected.txt

PASS window.atob("YQ==") is "a"
PASS window.atob("YWI=") is "ab"
PASS window.atob("YWJj") is "abc"
PASS window.atob("YWJjZA==") is "abcd"
PASS window.atob("YWJjZGU=") is "abcde"
PASS window.atob("YWJjZGVm") is "abcdef"
PASS window.btoa("a") is "YQ=="
PASS window.btoa("ab") is "YWI="
PASS window.btoa("abc") is "YWJj"
PASS window.btoa("abcd") is "YWJjZA=="
PASS window.btoa("abcde") is "YWJjZGU="
PASS window.btoa("abcdef") is "YWJjZGVm"
PASS typeof window.btoa is "function"
PASS window.btoa() threw exception TypeError: Failed to execute 'btoa' on 'Window': 1 argument required, but only 0 present..
PASS window.btoa("") is ""
PASS window.btoa(null) is "bnVsbA=="
PASS window.btoa(undefined) is "dW5kZWZpbmVk"
PASS window.btoa(window) is "W29iamVjdCBXaW5kb3dd"
PASS window.btoa("éé") is "6ek="
PASS window.btoa("\u0080\u0081") is "gIE="
PASS window.btoa("тест") threw exception InvalidCharacterError: Failed to execute 'btoa' on 'Window': The string to be encoded contains characters outside of the Latin1 range..
PASS window.btoa is 0
PASS typeof window.btoa is "number"
PASS typeof window.atob is "function"
PASS window.atob() threw exception TypeError: Failed to execute 'atob' on 'Window': 1 argument required, but only 0 present..
PASS window.atob("") is ""
PASS window.atob(null) is "ée"
PASS window.atob(undefined) threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob(" YQ==") is "a"
PASS window.atob("YQ== ") is "a"
PASS window.atob("YQ==    ") is "a"
PASS window.atob("YQ==\u000a") is "a"
PASS window.atob("YQ ==") is "a"
PASS window.atob("YQ\t==") is "a"
PASS window.atob("YQ\n==") is "a"
PASS window.atob("YQ\f==") is "a"
PASS window.atob("YQ\r==") is "a"
PASS window.atob("YQ\t\n\f\r ==") is "a"
PASS window.atob("6ek=") is "éé"
PASS window.atob("6ek") is "éé"
PASS window.atob("gIE=") is ""
PASS window.atob("тест") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded contains characters outside of the Latin1 range..
PASS window.atob("z") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("zz") is "Ï"
PASS window.atob("zzz") is "Ï<"
PASS window.atob("zzz=") is "Ï<"
PASS window.atob("zzz==") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("zzz===") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("zzz====") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("zzz=====") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("zzzz") is "Ï<ó"
PASS atob("            ") is ""
PASS window.atob("zzzzz") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("z=zz") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("=") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("==") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("===") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("====") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob("=====") threw exception InvalidCharacterError: Failed to execute 'atob' on 'Window': The string to be decoded is not correctly encoded..
PASS window.atob is 0
PASS typeof window.atob is "number"
PASS successfullyParsed is true

TEST COMPLETE
